feat: updated according to the Swagger Docu

+++ app/api/v1/schemas/exam.go
@@ -1,16 +1,16 @@
 package schemas
 
 type ExamRes struct {
-	Id          int     `json:"id"`
-	Description string  `json:"description"`
-	Weight      float64 `json:"weight"`
-	Type        string  `json:"string"`
+	Id          int    `json:"id"`
+	Description string `json:"description"`
+	Weight      int    `json:"weight"`
+	Type        string `json:"string"`
 }
 
 type ExamReq struct {
-	Description string  `json:"description"`
-	Weight      float64 `json:"weight"`
-	Type        string  `json:"type"`
+	Description string `json:"description"`
+	Weight      int    `json:"weight"`
+	Type        string `json:"type"`
 }
 
 type ExamReqStudent struct {
@@ -22,7 +22,7 @@ type ExamReqStudent struct {
 type CourseExamTeacher struct {
 	Id          int        `json:"id"`
 	Description string     `json:"description"`
-	Weight      float64    `json:"weight"`
+	Weight      int        `json:"weight"`
 	Type        string     `json:"type"`
 	AvgGrades   []GradType `json:"avgGrades"`
 }
@@ -30,7 +30,7 @@ type CourseExamTeacher struct {
 type CourseExamStudent struct {
 	Id          int      `json:"id"`
 	Description string   `json:"description"`
-	Weight      float64  `json:"weight"`
+	Weight      int      `json:"weight"`
 	Type        string   `json:"type"`
 	Grade       GradeRes `json:"grade"`
 }
